Moves ROOT LM output into GATESubscript. Modifies root file format.
This commit changes how root files are saved. Instead of everything in one file, the Coincidences and Singles are split into two: Sim_{TASK_ID}.Coincidences.root and Sim_{TASK_ID}.Singles.root. For my XCAT data (which has a ratio of approx 2:1:1 Trues:Scatter:Random events) the file size of the singles is just less than the coincidences. This is because a lot of the information in singles is not saved. (cannot further reduce size of coincidences).

With slight modification to the current unlisting setup, I should be able to unlist the coincidences. 
Following this we should be able to unlist singles for random correction.