Allow for optional template for norm file.
We need to perform the SSRB on the eff_factors (not the norm). Instead of using the "Classical" SSRB, we allow for the use of a template sinogram that the SSRB will match too, known as $norm_template. This will allow for the conversion between span-1 and span-N, where N is not limited to odd integers.