ExampleSTIR-GATE.sh to use uniform cylinder phantom with larger activity
The previous point source meant that there were no delayed events, which meant that the unlisting of the events would cause an error. This is fine, but in the Example (and main testing script) we do not want an error. Therefore, changing the phantom and increasing its activity will make it less likely to have 0 delayed events (~33).